+++ src/Types/Variant.php
@@ -4,15 +4,17 @@ declare(strict_types=1);
 
 namespace Gianfriaur\OpcuaPhpClient\Types;
 
-class Variant
+readonly class Variant
 {
     /**
      * @param BuiltinType $type
      * @param mixed $value
+     * @param null|int[] $dimensions
      */
     public function __construct(
-        private readonly BuiltinType $type,
-        private readonly mixed $value,
+        private BuiltinType $type,
+        private mixed       $value,
+        private ?array      $dimensions = null,
     ) {
     }
 
@@ -24,5 +26,18 @@ class Variant
     public function getValue(): mixed
     {
         return $this->value;
+    }
+
+    /**
+     * @return int[]|null
+     */
+    public function getDimensions(): ?array
+    {
+        return $this->dimensions;
+    }
+
+    public function isMultiDimensional(): bool
+    {
+        return $this->dimensions !== null && count($this->dimensions) > 1;
     }
 }